Implement support for an optional site-specific default configuration.
Gscope comes with a pre-configured user configuration for first time
users.  If those settings are not satisfactory, an optional 'site
specific' defaults override can be created.  This is most useful in a
multi-user installation.

Create a new 'default' settings file using Gscope [Options-->Startup
Preferences] and then copy that modified file to the same directory
where the Gscope binary is installed.  The default location for
saving user preferences is ~/.gscope/gscoperc